Triviolivia/requirements.txt
asgiref
channels==4.0.0
crispy-bootstrap5==0.7
daphne==4.0.0
Django==4.1.6
django-allauth
django-cors-headers
django-crispy-forms
django-extensions
django-filter
djangorestframework
lockfile==0.12.2
pandas
psycopg2
Pillow
pyOpenSSL==23.0.0
pytz==2022.1
ubuntu-advantage-tools
ubuntu-drivers-common

